cleaner solution for #13
move some of the logic to client side, by streaming a rolled up version
of twitter media and url entities combined, but ensmallifed and
intelligently ordered.

then on the client we do the actual replacement.  this lets us still
have a clean text field for the tweet.

should probably clean this up and move all of the tweet formatting logic
into a class now that more and more of this is moving client side, but
for now lets consider that this closes #13.